Handle migration tasks
