+++ client/package.json
@@ -22,7 +22,7 @@
     "@fullhuman/postcss-purgecss": "^4.1.3",
     "@playwright/test": "^1.25.1",
     "@sveltejs/adapter-node": "next",
-    "@sveltejs/kit": "^1.0.0-next.405",
+    "@sveltejs/kit": "^1.0.0-next.481",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/svelte": "^3.1.1",
     "@typescript-eslint/eslint-plugin": "^4.31.1",
