+++ client/src/app.html
@@ -15,10 +15,10 @@
     until SvelteKit adds the dynamic favicon link. -->
     <link rel="icon" href="data:," />
 
-    %svelte.head%
+    %sveltekit.head%
   </head>
 
   <body>
-    <div id="svelte">%svelte.body%</div>
+    <div>%sveltekit.body%</div>
   </body>
 </html>
