+++ client/src/routes/(app)/auth/datapass/create-organization/+layout.ts
@@ -1,0 +1,1 @@
+export const ssr = true;
