+++ client/src/routes/(app)/fiches/[id]/+page.ts
@@ -1,0 +1,16 @@
+import type { PageLoad } from "./$types";
+import { get } from "svelte/store";
+import { getDatasetByID } from "$lib/repositories/datasets";
+import { apiToken } from "$lib/stores/auth";
+
+export const load: PageLoad = async ({ fetch, params }) => {
+  const dataset = await getDatasetByID({
+    fetch,
+    apiToken: get(apiToken),
+    id: params.id,
+  });
+
+  return {
+    dataset,
+  };
+};
