+++ client/src/routes/(app)/fiches/[id]/_AsideItem.svelte
