+++ client/src/routes/(app)/fiches/[id]/edit/+layout@.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import BlankLayout from "$lib/layouts/BlankLayout.svelte";
+</script>
+
+<BlankLayout>
+  <slot />
+</BlankLayout>
