+++ client/src/routes/(app)/fiches/search/_FilterPanel.svelte
