+++ client/src/routes/(app)/login/+layout.ts
@@ -1,0 +1,1 @@
+export const ssr = true;
