+++ client/src/routes/+layout.ts
@@ -1,7 +1,28 @@
-import type { LayoutLoad.base } from "@sveltejs/kit";
-import { authGuard } from "$lib/auth/guard";
+import type { LayoutLoad } from "./$types";
+import { get } from "svelte/store";
+import { redirect } from "@sveltejs/kit";
+import { account } from "$lib/stores/auth";
+import { NON_AUTH_GUARDED_PAGES } from "src/constants";
+import { Maybe } from "$lib/util/maybe";
 
-export const load: LayoutLoad.base = async ({ url }) => {
-  throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292693)");
-  return authGuard(url);
+// NOTE: SSR usage is aimed at improving SEO, but most of our pages are
+// hidden behind authentication.
+// SSR will be enabled selectively on appropriate pages.
+// This also happens to simplify the e2e test setup, as auth state
+// for private pages can then be exclusively managed in the browser.
+// See: https://github.com/etalab/catalogage-donnees/pull/143
+export const ssr = false;
+
+export const load: LayoutLoad = async ({ url }) => {
+  // Force-redirect to the login page if an unauthenticated user
+  // is attempting to access a protected page.
+  if (NON_AUTH_GUARDED_PAGES.includes(url.pathname)) {
+    return {};
+  }
+
+  if (Maybe.Some(get(account))) {
+    return {};
+  }
+
+  throw redirect(302, "/");
 };
