--- client/src/routes/about/+page.svelte
@@ -1,48 +1,0 @@
-<script context="module">
-  throw new Error("@migration task: Check code was safely removed (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292722)");
-
-  // import { browser, dev } from "$app/env";
-
-  // // we don't need any JS on this page, though we'll load
-  // // it in dev so that we get hot module replacement...
-  // export const hydrate = dev;
-
-  // // ...but if the client-side router is already loaded
-  // // (i.e. we came here from elsewhere in the app), use it
-  // export const router = browser;
-
-  // // since there's no dynamic data here, we can prerender
-  // // it so that it gets served as a static asset in prod
-  // export const prerender = true;
-</script>
-
-<svelte:head>
-  <title>About</title>
-</svelte:head>
-
-<div class="content">
-  <h1>About this app</h1>
-
-  <p>
-    This is a <a href="https://kit.svelte.dev">SvelteKit</a> app. You can make your
-    own by typing the following into your command line and following the prompts:
-  </p>
-
-  <!-- TODO lose the @next! -->
-  <pre>npm init svelte@next</pre>
-
-  <p>
-    The page you're looking at is purely static HTML, with no client-side
-    interactivity needed. Because of that, we don't need to load any JavaScript.
-    Try viewing the page's source, or opening the devtools network panel and
-    reloading.
-  </p>
-</div>
-
-<style>
-  .content {
-    width: 100%;
-    max-width: var(--column-width);
-    margin: var(--column-margin-top) auto 0 auto;
-  }
-</style>
