--- client/src/routes/fiches/[id]/edit/+page@blank.svelte
@@ -1,188 +1,0 @@
-<script context="module" lang="ts">
-  throw new Error("@migration task: Check code was safely removed (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292722)");
-
-  // import type { Load } from "@sveltejs/kit";
-  // import { get } from "svelte/store";
-  // import { getCatalogBySiret } from "src/lib/repositories/catalogs";
-  // import { getDatasetByID, updateDataset } from "$lib/repositories/datasets";
-  // import type { Tag } from "src/definitions/tag";
-  // import { getTags } from "src/lib/repositories/tags";
-  // import { getLicenses } from "src/lib/repositories/licenses";
-  // import { getDatasetFiltersInfo } from "src/lib/repositories/datasetFilters";
-
-  // export const load: Load = async ({ fetch, params }) => {
-  //   const apiToken = get(apiTokenStore);
-  //   const siret = Maybe.expect(get(account), "$account").organizationSiret;
-
-  //   const [catalog, dataset, tags, licenses, filtersInfo] = await Promise.all([
-  //     getCatalogBySiret({ fetch, apiToken, siret }),
-  //     getDatasetByID({ fetch, apiToken, id: params.id }),
-  //     getTags({ fetch, apiToken }),
-  //     getLicenses({ fetch, apiToken }),
-  //     getDatasetFiltersInfo({ fetch, apiToken }),
-  //   ]);
-
-  //   return {
-  //     props: {
-  //       catalog,
-  //       dataset,
-  //       tags,
-  //       licenses,
-  //       filtersInfo,
-  //     },
-  //   };
-  // };
-</script>
-
-<script lang="ts">
-  throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
-
-  import { goto } from "$app/navigation";
-  import type { Catalog } from "src/definitions/catalogs";
-  import type { Dataset, DatasetFormData } from "src/definitions/datasets";
-  import DatasetForm from "$lib/components/DatasetForm/DatasetForm.svelte";
-  import paths from "$lib/paths";
-  import {
-    isAdmin,
-    apiToken as apiTokenStore,
-    account,
-  } from "$lib/stores/auth";
-  import { deleteDataset } from "$lib/repositories/datasets";
-  import { Maybe } from "$lib/util/maybe";
-  import DatasetFormLayout from "src/lib/components/DatasetFormLayout/DatasetFormLayout.svelte";
-  import ModalExitFormConfirmation from "src/lib/components/ModalExitFormConfirmation/ModalExitFormConfirmation.svelte";
-  import type { DatasetFiltersInfo } from "src/definitions/datasetFilters";
-
-  export let catalog: Maybe<Catalog>;
-  export let dataset: Maybe<Dataset>;
-  export let tags: Maybe<Tag[]>;
-  export let licenses: Maybe<string[]>;
-  export let filtersInfo: Maybe<DatasetFiltersInfo>;
-
-  let modalControlId = "stop-editing-form-modal";
-
-  let loading = false;
-
-  let formHasBeenTouched = false;
-
-  const history = window.history;
-
-  const onSave = async (event: CustomEvent<DatasetFormData>) => {
-    if (!Maybe.Some(dataset)) {
-      return;
-    }
-
-    const tagIds = event.detail.tags.map((item) => item.id);
-
-    try {
-      loading = true;
-
-      const updatedDataset = await updateDataset({
-        fetch,
-        apiToken: $apiTokenStore,
-        id: dataset.id,
-        data: { ...event.detail, tagIds },
-      });
-
-      if (Maybe.Some(updatedDataset)) {
-        await goto(paths.datasetDetail({ id: updatedDataset.id }));
-      }
-    } finally {
-      loading = false;
-    }
-  };
-
-  const onClickDelete = async (): Promise<void> => {
-    if (!Maybe.Some(dataset)) {
-      return;
-    }
-
-    const confirmed = confirm(
-      "Voulez-vous vraiment supprimer ce jeu de données ? Cette opération est irréversible."
-    );
-
-    if (!confirmed) {
-      return;
-    }
-
-    await deleteDataset({ fetch, apiToken: $apiTokenStore, id: dataset.id });
-    await goto(paths.home);
-  };
-
-  const handleExitForm = () => {
-    history.go(-1);
-  };
-</script>
-
-{#if Maybe.Some(catalog) && Maybe.Some(dataset) && Maybe.Some(tags) && Maybe.Some(licenses) && Maybe.Some(filtersInfo)}
-  <header class="fr-p-4w">
-    <h5>Modifier la fiche de jeu de données</h5>
-
-    {#if formHasBeenTouched}
-      <button
-        class="fr-btn fr-icon-close-line fr-btn--icon fr-btn--secondary"
-        data-fr-opened="false"
-        data-testid="exit-edit-form"
-        aria-controls={modalControlId}
-      >
-        {""}
-      </button>
-    {:else}
-      <button
-        data-testid="exit-edit-form"
-        class="fr-btn fr-icon-close-line fr-btn--icon fr-btn--secondary"
-        on:click={handleExitForm}
-      >
-        {""}
-      </button>
-    {/if}
-  </header>
-
-  <ModalExitFormConfirmation
-    on:confirm={handleExitForm}
-    controlId={modalControlId}
-  />
-
-  <DatasetFormLayout>
-    <DatasetForm
-      {catalog}
-      {tags}
-      {licenses}
-      geographicalCoverages={filtersInfo.geographicalCoverage}
-      initial={dataset}
-      {loading}
-      submitLabel="Enregistrer les modifications"
-      loadingLabel="Modification en cours..."
-      on:save={onSave}
-      on:touched={() => (formHasBeenTouched = true)}
-    />
-
-    {#if $isAdmin}
-      <div class="fr-alert fr-alert--error fr-mt-8w">
-        <p>
-          <strong> Zone de danger </strong>
-          <em>(visible car vous avez le rôle admin)</em>
-        </p>
-
-        <button
-          class="fr-btn fr-btn--secondary"
-          on:click|preventDefault={onClickDelete}
-        >
-          Supprimer ce jeu de données
-        </button>
-      </div>
-    {/if}
-  </DatasetFormLayout>
-{/if}
-
-<style>
-  header {
-    height: 10vh;
-    display: flex;
-    position: sticky;
-    justify-content: space-between;
-    top: 0;
-    z-index: 55;
-    background-color: var(--background-default-grey);
-  }
-</style>
