--- client/src/tests/app.env.mock.ts
@@ -1,4 +1,0 @@
-// Mock content for "$app/env" for testing purposes.
-export default {
-  browser: false,
-};
