+++ client/src/tests/app.environment.mock.ts
@@ -1,0 +1,4 @@
+// Mock content for "$app/environment" for testing purposes.
+export default {
+  browser: false,
+};
