+++ client/svelte.config.js
@@ -13,9 +13,9 @@ const config = {
     adapter: adapter(),
 
     // Override http methods in the Todo forms
-    methodOverride: {
-      allowed: ["PATCH", "DELETE"],
-    },
+    // methodOverride: {
+    //   allowed: ["PATCH", "DELETE"],
+    // },
 
     csp: {
       directives: {
