+++ client/vitest.config.ts
@@ -8,10 +8,10 @@ export default defineConfig({
   ...baseConfig,
   resolve: {
     alias: {
-      ...baseConfig.resolve.alias,
+      ...baseConfig?.resolve?.alias,
       // Add any alias resolutions that should be mocked, because
       // they are not available unless SvelteKit runs.
-      "$app/env": path.resolve("./src/tests/app.env.mock.ts"),
+      "$app/environment": path.resolve("./src/tests/app.env.mock.ts"),
       "$app/navigation": path.resolve("./src/tests/app.navigation.mock.ts"),
     },
   },
